nvim: Set up mini.splitjoin
mini.splitjoin is a small utility for splitting and joining lines
intelligently. It also supports local buffer-specific configs to modify
the behavior for specific filetypes. So for example in Swift, I want to
always add trailing commas for () and [] structures, and I want to pad
{} structures with spaces when collapsing them down to a single line. I
can configure this neatly inside the ftplugin.lua file (which I'm also
moving from a flat vim file)