nvim/sidekick: Set up sidekick
This lets me open an AI context in a sidebar via keybinding, and is
agent-agnostic so that I can use whatever agents I have available
without needing to change plugins every time they change.

DISCLAIMER: man do I not love that this is becoming a thing in our
field. Seriously, adding this feels like admitting defeat in a lot of
ways. However, the pressure to adopt these kinds of tools is absolutely
overwhelming and it starts to feel like fighting against a riptide. I
personally still believe that these tools are _extremely_ questionable
ethically and economically and I have serious concerns about the long
term impact these things will have on all of our careers. But for _now_
I'm resigned to using them. I do think they are actually good as
educational tools (that is: "how do I do X in language Y" is a faster
mode of discovery than searching online and filtering through shitty
results), and they _can_ be effective aids for specific kinds of tasks
when used with purpose.