nvim/lsp: Add copilot LSP support
See general AI disclaimer in 179b7df1748a

Note that the inline completion stuff is there but commented out.
Support exists in the nightly builds, but not in any official release.
Once it does I will be able to uncomment that code (which comes from the
lspconfig documentation for the copilot LSP) and get fancy inline
autocomplete for larger code suggestions.