nvim/mini: Replace some config with mini.basics
mini.basics sets a bunch of useful config out of the box, and even sets
up some useful mappings and autocommands. This actually removes the need
for vim-system-copy, since it comes with simple keybindings for that. It
also removes an autocommand I had set up for highlighting on yank.

Interestingly, the listchars config seems broken for me for reasons I
truly don't understand. So even though _technically_ it's duplicated
config now, I'm going to keep it around.