nvim: Simplify lsp setup
Ok this did get away from me a bit, but the _intent_ was to simplify the
giant lsp config so that I actually understood what was going on in
there. As a result I ended up breaking out things like keymaps and
autocommands into their own files, and then that led to breaking out
_more_ things into their own files, and so here we are. However, I do
think everything is _much_ clearer now, since I dropped a whole bunch of
what seemed to be unnecessary capability configuration stuff in favor of
a very straightforward setup. It's possible some things don't work the
way they used to now but it _seems_ like things are working properly so
I'll just keep moving forward and circle back to this if I need to do
something different later.