nvim: Enable exrc
This has gotten to be really great since nvim 0.9. Before then, enabling
this was just introducing a serious security vulnerability for yourself
since vim would blindly load any local config that it found. But now it
uses a trust pattern based on a hash of the config being loaded. So any
time it changes you're given an opportunity to ignore it, load it, or
view it. This is excellent and having this functionality available is
going to be really nice!