nvim/scooter: Move scooter integration into external-tui config
With the newest version of Scooter, I can configure the editor command
from the CLI, which means it works with my external-tui plugin. This
also makes Scooter more useful outside of neovim since we can now use
the default behavior (which uses $EDITOR). This was broken in the old
config since there wasn't any neovim instance available to attach to.